[TASK] De-couple active extensions from .env
Although it is great to read the active extensions
from the environment, in most cases they are set in stone
for a specific project.

Reflect that by adding a composer script that set the env var there accordingly.

We can now generate the package state simply by calling:

composer run package-states